Working functions
LD2Struct, Add2Struct and PSearch functions all seem to work.

Known issue: variables "valid" and "yesNo" should only handle a single char. But if user enters an array of chars, a looping effect occurs until the end of the char array.

Need to figure out how to handle that.